I made a windows service that is supposed to run a server that listens for clients and then receives data that will go into the mouse_event() function to move the mouse. The service is supposed to put int the registry information about the server's port and ip in the LAN.
The service still doesn't update registry for ip and port, requires chages. I also added a Windows form application that makes an icon appear in the system tray(all these icons in the bottom right corner of the screen). Clicking on the icon will display a qr code that is supposed to contain ip and port values that it found in the registry(however the service wouldn't update the registry properly).